relay/requirements.txt
# Mandali Teams Relay Service Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
websockets>=12.0
botbuilder-core>=4.16.0
botbuilder-integration-aiohttp>=4.16.0
aiohttp>=3.9.0
